arch-riscv,cpu-o3: Checkpoint current matrix metadata refactor
