+++ src/cpu/o3/iew.hh
@@ -56,9 +56,7 @@
 #include "cpu/o3/inst_queue.hh"
 #include "cpu/o3/limits.hh"
 #include "cpu/o3/lsq.hh"
-#include "cpu/o3/mls_replay_queue.hh"
 #include "cpu/o3/mls_unit.hh"
-#include "cpu/o3/mls_virtual_queue.hh"
 #include "cpu/o3/rob.hh"
 #include "cpu/o3/scoreboard.hh"
 #include "cpu/timebuf.hh"
