+++ src/cpu/o3/lsq.hh
@@ -64,9 +64,7 @@
 #include "cpu/inst_seq.hh"
 #include "cpu/o3/dyn_inst_ptr.hh"
 #include "cpu/o3/dyn_inst_xsmeta.hh"
-#include "cpu/o3/mls_replay_queue.hh"
 #include "cpu/o3/mls_unit.hh"
-#include "cpu/o3/mls_virtual_queue.hh"
 #include "cpu/utils.hh"
 #include "enums/SMTQueuePolicy.hh"
 #include "mem/packet.hh"
