+++ src/matrix/MemoryAdapter.cc
@@ -26,7 +26,8 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "matrix/matrix_memory_adapter.hh"
+#include "matrix/MemoryAdapter.hh"
+
 #include "mem/port_proxy.hh"
 #include "mem/se_translating_port_proxy.hh"
 #include "mem/translating_port_proxy.hh"
@@ -104,6 +105,84 @@ writeMatrixElem(const AmuLsuDesc &desc, PortProxy &fallback, Addr addr, T value)
 }
 
 } // anonymous namespace
+
+bool
+NullMatrixMemoryAdapter::loadTile(const AmuLsuDesc &desc, MatrixTensor &out_tensor)
+{
+    (void)desc;
+    (void)out_tensor;
+    return false;
+}
+
+bool
+NullMatrixMemoryAdapter::storeTile(const AmuLsuDesc &desc,
+                                   const MatrixTensor &tensor)
+{
+    (void)desc;
+    (void)tensor;
+    return false;
+}
+
+bool
+SparseMatrixMemoryAdapter::loadTile(const AmuLsuDesc &desc,
+                                    MatrixTensor &out_tensor)
+{
+    out_tensor.rows = desc.row;
+    out_tensor.cols = desc.column;
+    out_tensor.elemType = desc.elemType;
+    out_tensor.elements.clear();
+    out_tensor.elements.reserve(static_cast<size_t>(desc.row) * desc.column);
+
+    for (uint32_t r = 0; r < desc.row; ++r) {
+        for (uint32_t c = 0; c < desc.column; ++c) {
+            int64_t value = 0;
+            auto it = elements.find(elemAddr(desc, r, c));
+            if (it != elements.end()) {
+                value = it->second;
+            }
+            out_tensor.elements.push_back(value);
+        }
+    }
+
+    return true;
+}
+
+bool
+SparseMatrixMemoryAdapter::storeTile(const AmuLsuDesc &desc,
+                                     const MatrixTensor &tensor)
+{
+    if (tensor.rows != desc.row || tensor.cols != desc.column ||
+        tensor.elemType != desc.elemType) {
+        return false;
+    }
+
+    for (uint32_t r = 0; r < desc.row; ++r) {
+        for (uint32_t c = 0; c < desc.column; ++c) {
+            elements[elemAddr(desc, r, c)] =
+                tensor.elements[static_cast<size_t>(r) * tensor.cols + c];
+        }
+    }
+
+    return true;
+}
+
+void
+SparseMatrixMemoryAdapter::writeElement(Addr addr, int64_t value)
+{
+    elements[addr] = value;
+}
+
+bool
+SparseMatrixMemoryAdapter::readElement(Addr addr, int64_t &value) const
+{
+    auto it = elements.find(addr);
+    if (it == elements.end()) {
+        return false;
+    }
+
+    value = it->second;
+    return true;
+}
 
 Gem5MatrixMemoryAdapter::Gem5MatrixMemoryAdapter(PortProxy &proxy)
   : portProxy(&proxy)
