+++ src/matrix/MemoryLoader.cc
@@ -28,7 +28,7 @@
 
 #include <utility>
 
-#include "matrix/detailed_cute_backend.hh"
+#include "matrix/CUTETOP.hh"
 
 namespace gem5
 {
