+++ src/matrix/SConscript
@@ -5,27 +5,22 @@ Import('*')
 DebugFlag('MatrixCuteTrace')
 
 Source('matrix_regfile.cc')
-Source('matrix_reg_resource.cc')
-Source('matrix_memory_adapter.cc')
-Source('matrix_memory_adapter_gem5.cc')
-Source('taskcontrol.cc')
-Source('backend_runtime.cc')
-Source('detailed_cute_scoreboard.cc')
-Source('mregfile.cc')
-Source('memoryload.cc')
-Source('matrix_ex.cc')
+Source('MemoryAdapter.cc')
+Source('TaskController.cc')
+Source('CUTETOP.cc')
+Source('Scoreboard.cc')
+Source('MemoryLoader.cc')
+Source('MatrixTE.cc')
 
 GTest(
-    'detailed_cute_backend.test',
-    'detailed_cute_backend.test.cc',
+    'CUTETOP.test',
+    'CUTETOP.test.cc',
     'matrix_regfile.cc',
-    'matrix_reg_resource.cc',
-    'matrix_memory_adapter.cc',
-    'detailed_cute_scoreboard.cc',
-    'taskcontrol.cc',
-    'backend_runtime.cc',
-    'mregfile.cc',
-    'memoryload.cc',
-    'matrix_ex.cc',
+    'MemoryAdapter.cc',
+    'Scoreboard.cc',
+    'TaskController.cc',
+    'CUTETOP.cc',
+    'MemoryLoader.cc',
+    'MatrixTE.cc',
     with_tag('gem5 trace'),
 )
